Moved solver to new class
