+++ Assets/Resources/Scripts/BoardSolver.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: f714a36e6c156415196786410c041b79
+timeCreated: 1457896122
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
